General cleanup of files after formatter setting change.
Changed formatter so that empty lines are properly indented instead of
being automatically reduced to column zero. Following this change, the
new formatting style was applied to all of the source files.

During this process, a unified source file header was applied to each
and every source file, and care was taken to ensure that contributors
for each of the files was preserved.

Also removed duplicate of source file "IntegerEditor". The removed copy
was in the jsea package of FHAES, and the kept copy remains in
segmentation package of FHUtil.

Finally, updated javadocs to reflect the mass changes and added a copy
of the formatter settings file under FHAES/Formatter in case others wish
to use it while coding.