Handled some concurrency issues with the FeedbackMessagePanel.
Also fixed FHRecorder so that you can close and/or discard a new file
without being prompted to save; this is only in the case that no
modifications have been made since the new file was opened in
FHRecorder.

Also fixed a bug with reloading category entries after a file has been
edited in FHRecorder. If auto-loading for categories is set to occur,
the category file will be re-attached to an FHX file after modifications
made to that FHX file in FHRecorder are saved.

Also added two new feedback messages: one which tells the user that
categories have been saved, and another which informs the user that an
invalid FHX file cannot have it's categories modified.