Finished working on all issues regarding the BCADYearSpinner.
Installed BCADYearSpinners on the NewSampleDialog and SampleInputPanel.
Also reworked most of the internal code for NewSampleDialog so that it
is cleaner and functions in a more straight-forward manner. Also refined
the layout of the NewSampleDialog.

Finally, went back to some previous work areas and made sure locally
defined classes were marked as private.