Began refactoring internal architecture of FireChartSVG.
Instead of making a "util" package in "org.fhaes.neofhchart" there will
instead be a package called "svg" which will contain the FireChartSVG
class and all of its helper classes as they are created. The goal is to
reduce the complexity of FireChartSVG by breaking it down into smaller
parts which contain related functionality.