Added support for colorizing category groups and labels. Addresses #93.
If selected, colors are automatically picked for the groups and labels.
This was implemented as such to avoid the need for making a complex
window for selecting colors for each group in the chart. Automatically
picking the colors is also more efficient in the case that an FHX file
contains a large number of series with different categories.